+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/wallet/lib/data/CredentialSubject.kt
@@ -4,7 +4,8 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 /**
- * Base class for the subject of a [VerifiableCredential], see subclasses of this class, e.g. [AtomicAttribute2023].
+ * Base class for the subject of a [VerifiableCredential], see subclasses of this class, e.g., a concrete credential
+ * implementation
  */
 @Serializable
 abstract class CredentialSubject {